+++ DSCResources/MSFT_SqlSetup/MSFT_SqlSetup.psm1
@@ -293,6 +293,17 @@ function Get-TargetResource
         $analysisLogDirectory = $analysisServer.ServerProperties['LogDir'].Value
         $analysisBackupDirectory = $analysisServer.ServerProperties['BackupDir'].Value
 
+        <#
+            The property $analysisServer.ServerMode.value__ contains the
+            server mode (aka deployment mode) value 0, 1 or 2. See DeploymentMode
+            here https://docs.microsoft.com/en-us/sql/analysis-services/server-properties/general-properties.
+
+            The property $analysisServer.ServerMode contains the display name of
+            the property value__. See more information here
+            https://msdn.microsoft.com/en-us/library/microsoft.analysisservices.core.server.servermode.aspx.
+        #>
+        $analysisServerMode = $analysisServer.ServerMode.ToString().ToUpper()
+
         $analysisSystemAdminAccounts = [System.String[]] $analysisServer.Roles['Administrators'].Members.Name
 
         $analysisConfigDirectory = (Get-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\$analysisServiceName" -Name 'ImagePath').ImagePath.Replace(' -s ',',').Split(',')[1].Trim('"')
@@ -487,6 +498,7 @@ function Get-TargetResource
         ASBackupDir = $analysisBackupDirectory
         ASTempDir = $analysisTempDirectory
         ASConfigDir = $analysisConfigDirectory
+        ASServerMode = $analysisServerMode
         ISSvcAccountUsername = $integrationServiceAccountUsername
         FailoverClusterGroupName = $clusteredSqlGroupName
         FailoverClusterNetworkName = $clusteredSqlHostname
@@ -618,6 +630,13 @@ function Get-TargetResource
 
     .PARAMETER ASConfigDir
         Path for Analysis Services config.
+
+    .PARAMETER ASServerMode
+        The server mode for SQL Server Analysis Services instance. The default is
+        to install in Multidimensional mode. Valid values in a cluster scenario
+        are MULTIDIMENSIONAL or TABULAR. Parameter ASServerMode is case-sensitive.
+        All values must be expressed in upper case.
+        { MULTIDIMENSIONAL | TABULAR | POWERPIVOT }.
 
     .PARAMETER ISSvcAccount
        Service account for Integration Services service.
@@ -796,6 +815,11 @@ function Set-TargetResource
         [Parameter()]
         [System.String]
         $ASConfigDir,
+
+        [Parameter()]
+        [ValidateSet('MULTIDIMENSIONAL','TABULAR','POWERPIVOT', IgnoreCase = $false)]
+        [System.String]
+        $ASServerMode,
 
         [Parameter()]
         [System.Management.Automation.PSCredential]
@@ -1244,6 +1268,12 @@ function Set-TargetResource
             'ASConfigDir'
         )
 
+
+        if ($PSBoundParameters.ContainsKey('ASServerMode'))
+        {
+            $setupArguments['ASServerMode'] = $ASServerMode
+        }
+
         if ($PSBoundParameters.ContainsKey('ASSvcAccount'))
         {
             $setupArguments += (Get-ServiceAccountParameters -ServiceAccount $ASSvcAccount -ServiceType 'AS')
@@ -1540,6 +1570,13 @@ function Set-TargetResource
     .PARAMETER ASConfigDir
         Path for Analysis Services config.
 
+    .PARAMETER ASServerMode
+        The server mode for SQL Server Analysis Services instance. The default is
+        to install in Multidimensional mode. Valid values in a cluster scenario
+        are MULTIDIMENSIONAL or TABULAR. Parameter ASServerMode is case-sensitive.
+        All values must be expressed in upper case.
+        { MULTIDIMENSIONAL | TABULAR | POWERPIVOT }.
+
     .PARAMETER ISSvcAccount
        Service account for Integration Services service.
 
@@ -1716,6 +1753,11 @@ function Test-TargetResource
         [Parameter()]
         [System.String]
         $ASConfigDir,
+
+        [Parameter()]
+        [ValidateSet('MULTIDIMENSIONAL','TABULAR','POWERPIVOT', IgnoreCase = $false)]
+        [System.String]
+        $ASServerMode,
 
         [Parameter()]
         [System.Management.Automation.PSCredential]