+++ Tests/Integration/MSFT_SqlSetup.Integration.Tests.ps1
@@ -47,9 +47,13 @@ try
     . $configFile
 
     # These sets variables used for verification from the dot-sourced $ConfigurationData variable.
-    $mockInstanceName = $ConfigurationData.AllNodes.InstanceName
-    $mockFeatures = $ConfigurationData.AllNodes.Features
-    $mockSqlCollation = $ConfigurationData.AllNodes.SQLCollation
+    $mockSqlEngineInstanceName = $ConfigurationData.AllNodes.SqlEngineInstanceName
+    $mockSqlEngineFeatures = $ConfigurationData.AllNodes.SqlEngineFeatures
+    $mockAnalysisServicesTabularInstanceName = $ConfigurationData.AllNodes.AnalysisServicesTabularInstanceName
+    $mockAnalysisServicesTabularFeatures = $ConfigurationData.AllNodes.AnalysisServicesTabularFeatures
+    $mockAnalysisServicesTabularServerMode = $ConfigurationData.AllNodes.AnalysisServicesTabularServerMode
+    $mockAnalysisServicesMultiServerMode = $ConfigurationData.AllNodes.AnalysisServicesMultiServerMode
+    $mockCollation = $ConfigurationData.AllNodes.Collation
     $mockInstallSharedDir = $ConfigurationData.AllNodes.InstallSharedDir
     $mockInstallSharedWOWDir = $ConfigurationData.AllNodes.InstallSharedWOWDir
     $mockUpdateEnable = $ConfigurationData.AllNodes.UpdateEnabled
@@ -99,10 +103,12 @@ try
     $mockSqlAgentServiceAccountUserName = "$env:COMPUTERNAME\svc-SqlAgent"
     $mockSqlAgentServiceCredential = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList $mockSqlAgentServiceAccountUserName, $mockSqlAgentServiceAccountPassword
 
-    $configurationName = "$($script:DSCResourceName)_InstallSqlEngineAsSystem_Config"
-    $resourceId = "[$($script:DSCResourceFriendlyName)]Integration_Test"
+    Describe "$($script:DSCResourceName)_InstallSqlEngineAsSystem_Integration" {
+        BeforeAll {
+            $configurationName = "$($script:DSCResourceName)_InstallSqlEngineAsSystem_Config"
+            $resourceId = "[$($script:DSCResourceFriendlyName)]Integration_Test"
+        }
 
-    Describe "$($script:DSCResourceName)_Integration" {
         It 'Should compile and apply the MOF without throwing' {
             {
                 $configurationParameters = @{
@@ -173,30 +179,35 @@ try
             $resourceCurrentState.Action                     | Should -BeNullOrEmpty
             $resourceCurrentState.AgtSvcAccount              | Should -BeNullOrEmpty
             $resourceCurrentState.AgtSvcAccountUsername      | Should -Be ('.\{0}' -f (Split-Path -Path $mockSqlAgentServiceAccountUserName -Leaf))
-            $resourceCurrentState.ASBackupDir                | Should -BeNullOrEmpty
-            $resourceCurrentState.ASCollation                | Should -BeNullOrEmpty
-            $resourceCurrentState.ASConfigDir                | Should -BeNullOrEmpty
-            $resourceCurrentState.ASDataDir                  | Should -BeNullOrEmpty
-            $resourceCurrentState.ASLogDir                   | Should -BeNullOrEmpty
+            $resourceCurrentState.ASServerMode               | Should -Be $mockAnalysisServicesMultiServerMode
+            $resourceCurrentState.ASBackupDir                | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockSqlEngineInstanceName\OLAP\Backup")
+            $resourceCurrentState.ASCollation                | Should -Be $mockCollation
+            $resourceCurrentState.ASConfigDir                | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockSqlEngineInstanceName\OLAP\Config")
+            $resourceCurrentState.ASDataDir                  | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockSqlEngineInstanceName\OLAP\Data")
+            $resourceCurrentState.ASLogDir                   | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockSqlEngineInstanceName\OLAP\Log")
+            $resourceCurrentState.ASTempDir                  | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockSqlEngineInstanceName\OLAP\Temp")
             $resourceCurrentState.ASSvcAccount               | Should -BeNullOrEmpty
-            $resourceCurrentState.ASSvcAccountUsername       | Should -BeNullOrEmpty
-            $resourceCurrentState.ASSysAdminAccounts         | Should -BeNullOrEmpty
-            $resourceCurrentState.ASTempDir                  | Should -BeNullOrEmpty
+            $resourceCurrentState.ASSvcAccountUsername       | Should -Be ('.\{0}' -f (Split-Path -Path $mockSqlServiceAccountUserName -Leaf))
+            $resourceCurrentState.ASSysAdminAccounts         | Should -Be @(
+                $mockSqlAdminAccountUserName,
+                $mockSqlInstallAccountUserName,
+                "NT SERVICE\SSASTELEMETRY`$$mockSqlEngineInstanceName"
+            )
             $resourceCurrentState.BrowserSvcStartupType      | Should -BeNullOrEmpty
             $resourceCurrentState.ErrorReporting             | Should -BeNullOrEmpty
             $resourceCurrentState.FailoverClusterGroupName   | Should -BeNullOrEmpty
             $resourceCurrentState.FailoverClusterIPAddress   | Should -BeNullOrEmpty
             $resourceCurrentState.FailoverClusterNetworkName | Should -BeNullOrEmpty
-            $resourceCurrentState.Features                   | Should -Be $mockFeatures
+            $resourceCurrentState.Features                   | Should -Be $mockSqlEngineFeatures
             $resourceCurrentState.ForceReboot                | Should -BeNullOrEmpty
             $resourceCurrentState.FTSvcAccount               | Should -BeNullOrEmpty
             $resourceCurrentState.FTSvcAccountUsername       | Should -BeNullOrEmpty
             $resourceCurrentState.InstallSharedDir           | Should -Be $mockInstallSharedDir
             $resourceCurrentState.InstallSharedWOWDir        | Should -Be $mockInstallSharedWOWDir
-            $resourceCurrentState.InstallSQLDataDir          | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSSQL13.$mockInstanceName\MSSQL")
+            $resourceCurrentState.InstallSQLDataDir          | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSSQL13.$mockSqlEngineInstanceName\MSSQL")
             $resourceCurrentState.InstanceDir                | Should -Be $mockInstallSharedDir
-            $resourceCurrentState.InstanceID                 | Should -Be $mockInstanceName
-            $resourceCurrentState.InstanceName               | Should -Be $mockInstanceName
+            $resourceCurrentState.InstanceID                 | Should -Be $mockSqlEngineInstanceName
+            $resourceCurrentState.InstanceName               | Should -Be $mockSqlEngineInstanceName
             $resourceCurrentState.ISSvcAccount               | Should -BeNullOrEmpty
             $resourceCurrentState.ISSvcAccountUsername       | Should -BeNullOrEmpty
             $resourceCurrentState.ProductKey                 | Should -BeNullOrEmpty
@@ -207,23 +218,154 @@ try
             $resourceCurrentState.SetupProcessTimeout        | Should -BeNullOrEmpty
             $resourceCurrentState.SourceCredential           | Should -BeNullOrEmpty
             $resourceCurrentState.SourcePath                 | Should -Be "$($mockIsoMediaDriveLetter):\"
-            $resourceCurrentState.SQLBackupDir               | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSSQL13.$mockInstanceName\MSSQL\Backup")
-            $resourceCurrentState.SQLCollation               | Should -Be $mockSqlCollation
+            $resourceCurrentState.SQLBackupDir               | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSSQL13.$mockSqlEngineInstanceName\MSSQL\Backup")
+            $resourceCurrentState.SQLCollation               | Should -Be $mockCollation
             $resourceCurrentState.SQLSvcAccount              | Should -BeNullOrEmpty
             $resourceCurrentState.SQLSvcAccountUsername      | Should -Be ('.\{0}' -f (Split-Path -Path $mockSqlServiceAccountUserName -Leaf))
             $resourceCurrentState.SQLSysAdminAccounts        | Should -Be @(
                 $mockSqlAdminAccountUserName,
                 $mockSqlInstallAccountUserName,
-                "NT SERVICE\MSSQL`$$mockInstanceName",
-                "NT SERVICE\SQLAgent`$$mockInstanceName",
+                "NT SERVICE\MSSQL`$$mockSqlEngineInstanceName",
+                "NT SERVICE\SQLAgent`$$mockSqlEngineInstanceName",
                 'NT SERVICE\SQLWriter',
                 'NT SERVICE\Winmgmt',
                 'sa'
             )
             $resourceCurrentState.SQLTempDBDir               | Should -BeNullOrEmpty
             $resourceCurrentState.SQLTempDBLogDir            | Should -BeNullOrEmpty
-            $resourceCurrentState.SQLUserDBDir               | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSSQL13.$mockInstanceName\MSSQL\DATA\")
-            $resourceCurrentState.SQLUserDBLogDir            | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSSQL13.$mockInstanceName\MSSQL\DATA\")
+            $resourceCurrentState.SQLUserDBDir               | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSSQL13.$mockSqlEngineInstanceName\MSSQL\DATA\")
+            $resourceCurrentState.SQLUserDBLogDir            | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSSQL13.$mockSqlEngineInstanceName\MSSQL\DATA\")
+            $resourceCurrentState.SQMReporting               | Should -BeNullOrEmpty
+            $resourceCurrentState.SuppressReboot             | Should -BeNullOrEmpty
+            $resourceCurrentState.UpdateEnabled              | Should -BeNullOrEmpty
+            $resourceCurrentState.UpdateSource               | Should -BeNullOrEmpty
+
+        }
+    }
+
+    Describe "$($script:DSCResourceName)_InstallAnalysisServicesAsSystem_Integration" {
+        BeforeAll {
+            $configurationName = "$($script:DSCResourceName)_InstallAnalysisServicesAsSystem_Config"
+            $resourceId = "[$($script:DSCResourceFriendlyName)]Integration_Test"
+        }
+
+        It 'Should compile and apply the MOF without throwing' {
+            {
+                $configurationParameters = @{
+                    SqlInstallCredential = $mockSqlInstallCredential
+                    SqlAdministratorCredential = $mockSqlAdminCredential
+                    SqlServiceCredential = $mockSqlServiceCredential
+                    OutputPath = $TestDrive
+                    # The variable $ConfigurationData was dot-sourced above.
+                    ConfigurationData = $ConfigurationData
+                }
+
+                & $configurationName @configurationParameters
+
+                $startDscConfigurationParameters = @{
+                    Path = $TestDrive
+                    ComputerName = 'localhost'
+                    Wait = $true
+                    Verbose = $true
+                    Force = $true
+                    ErrorAction = 'Stop'
+                }
+
+                Start-DscConfiguration @startDscConfigurationParameters
+            } | Should -Not -Throw
+        } -ErrorVariable itBlockError
+
+        # Check if previous It-block failed. If so output the SQL Server setup log file.
+        if ( $itBlockError.Count -ne 0 )
+        {
+            <#
+                Below code will output the Summary.txt log file, this is to be
+                able to debug any problems that potentially occurred during setup.
+                This will pick up the newest Summary.txt log file, so any
+                other log files will be ignored (AppVeyor build worker has
+                SQL Server instances installed by default).
+                This code is meant to work regardless what SQL Server
+                major version is used for the integration test.
+            #>
+            $summaryLogPath = Get-ChildItem -Path 'C:\Program Files\Microsoft SQL Server\**\Setup Bootstrap\Log\Summary.txt' |
+                Sort-Object -Property LastWriteTime -Descending |
+                Select-Object -First 1
+
+            $summaryLog = Get-Content $summaryLogPath
+
+            Write-Verbose -Message $('-' * 80) -Verbose
+            Write-Verbose -Message 'Summary.txt' -Verbose
+            Write-Verbose -Message $('-' * 80) -Verbose
+            $summaryLog | ForEach-Object {
+                Write-Verbose $_ -Verbose
+            }
+            Write-Verbose -Message $('-' * 80) -Verbose
+        }
+
+        It 'Should be able to call Get-DscConfiguration without throwing' {
+            { Get-DscConfiguration -Verbose -ErrorAction Stop } | Should -Not -Throw
+        }
+
+        It 'Should have set the resource and all the parameters should match' {
+            $currentConfiguration = Get-DscConfiguration
+
+            $resourceCurrentState = $currentConfiguration | Where-Object -FilterScript {
+                $_.ConfigurationName -eq $configurationName
+            } | Where-Object -FilterScript {
+                $_.ResourceId -eq $resourceId
+            }
+
+            $resourceCurrentState.Action                     | Should -BeNullOrEmpty
+            $resourceCurrentState.AgtSvcAccount              | Should -BeNullOrEmpty
+            $resourceCurrentState.AgtSvcAccountUsername      | Should -BeNullOrEmpty
+            $resourceCurrentState.ASServerMode               | Should -Be $mockAnalysisServicesTabularServerMode
+            $resourceCurrentState.ASBackupDir                | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockAnalysisServicesTabularInstanceName\OLAP\Backup")
+            $resourceCurrentState.ASCollation                | Should -Be $mockCollation
+            $resourceCurrentState.ASConfigDir                | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockAnalysisServicesTabularInstanceName\OLAP\Config")
+            $resourceCurrentState.ASDataDir                  | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockAnalysisServicesTabularInstanceName\OLAP\Data")
+            $resourceCurrentState.ASLogDir                   | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockAnalysisServicesTabularInstanceName\OLAP\Log")
+            $resourceCurrentState.ASTempDir                  | Should -Be (Join-Path -Path $mockInstallSharedDir -ChildPath "MSAS13.$mockAnalysisServicesTabularInstanceName\OLAP\Temp")
+            $resourceCurrentState.ASSvcAccount               | Should -BeNullOrEmpty
+            $resourceCurrentState.ASSvcAccountUsername       | Should -Be ('.\{0}' -f (Split-Path -Path $mockSqlServiceAccountUserName -Leaf))
+            $resourceCurrentState.ASSysAdminAccounts         | Should -Be @(
+                $mockSqlAdminAccountUserName,
+                $mockSqlInstallAccountUserName,
+                "NT SERVICE\SSASTELEMETRY`$$mockAnalysisServicesTabularInstanceName"
+            )
+            $resourceCurrentState.BrowserSvcStartupType      | Should -BeNullOrEmpty
+            $resourceCurrentState.ErrorReporting             | Should -BeNullOrEmpty
+            $resourceCurrentState.FailoverClusterGroupName   | Should -BeNullOrEmpty
+            $resourceCurrentState.FailoverClusterIPAddress   | Should -BeNullOrEmpty
+            $resourceCurrentState.FailoverClusterNetworkName | Should -BeNullOrEmpty
+            $resourceCurrentState.Features                   | Should -Be $mockAnalysisServicesTabularFeatures
+            $resourceCurrentState.ForceReboot                | Should -BeNullOrEmpty
+            $resourceCurrentState.FTSvcAccount               | Should -BeNullOrEmpty
+            $resourceCurrentState.FTSvcAccountUsername       | Should -BeNullOrEmpty
+            $resourceCurrentState.InstallSharedDir           | Should -Be $mockInstallSharedDir
+            $resourceCurrentState.InstallSharedWOWDir        | Should -Be $mockInstallSharedWOWDir
+            $resourceCurrentState.InstallSQLDataDir          | Should -BeNullOrEmpty
+            $resourceCurrentState.InstanceDir                | Should -BeNullOrEmpty
+            $resourceCurrentState.InstanceID                 | Should -BeNullOrEmpty
+            $resourceCurrentState.InstanceName               | Should -Be $mockAnalysisServicesTabularInstanceName
+            $resourceCurrentState.ISSvcAccount               | Should -BeNullOrEmpty
+            $resourceCurrentState.ISSvcAccountUsername       | Should -BeNullOrEmpty
+            $resourceCurrentState.ProductKey                 | Should -BeNullOrEmpty
+            $resourceCurrentState.RSSvcAccount               | Should -BeNullOrEmpty
+            $resourceCurrentState.RSSvcAccountUsername       | Should -BeNullOrEmpty
+            $resourceCurrentState.SAPwd                      | Should -BeNullOrEmpty
+            $resourceCurrentState.SecurityMode               | Should -BeNullOrEmpty
+            $resourceCurrentState.SetupProcessTimeout        | Should -BeNullOrEmpty
+            $resourceCurrentState.SourceCredential           | Should -BeNullOrEmpty
+            $resourceCurrentState.SourcePath                 | Should -Be "$($mockIsoMediaDriveLetter):\"
+            $resourceCurrentState.SQLBackupDir               | Should -BeNullOrEmpty
+            $resourceCurrentState.SQLCollation               | Should -BeNullOrEmpty
+            $resourceCurrentState.SQLSvcAccount              | Should -BeNullOrEmpty
+            $resourceCurrentState.SQLSvcAccountUsername      | Should -BeNullOrEmpty
+            $resourceCurrentState.SQLSysAdminAccounts        | Should -BeNullOrEmpty
+            $resourceCurrentState.SQLTempDBDir               | Should -BeNullOrEmpty
+            $resourceCurrentState.SQLTempDBLogDir            | Should -BeNullOrEmpty
+            $resourceCurrentState.SQLUserDBDir               | Should -BeNullOrEmpty
+            $resourceCurrentState.SQLUserDBLogDir            | Should -BeNullOrEmpty
             $resourceCurrentState.SQMReporting               | Should -BeNullOrEmpty
             $resourceCurrentState.SuppressReboot             | Should -BeNullOrEmpty
             $resourceCurrentState.UpdateEnabled              | Should -BeNullOrEmpty